Add DB/Candidate fields to store subfolder links.
Change candidate main folder creation so that it does not go looking for folders on drive if a link is alreday stored in the DB.

Automatically create candidate subfolders when a candidate folder is created.
Also store links to those subfolders in the DB if they are not already present.

Added methods to CandidateService around subfolders and using the CandidateSubfolderType.

Removed addFakeTasks methods - no longer needed

Remove any null string folderlinks - prevent creation of new null string folderlinks