Define "Live Jobs" as needing to be published.
Also change definition of "accepting" in the search request to be based on stage.

Stage filter in live jobs, overrides the new "accepting" which also depends on stage - you can't have them both filtering.

A job's stage can't be edited if the job has not been published.

Remove use of job's accepting field