Remove option for users being able to associate a job with a list.
Adding candidates to a list only creates cases if the list is a submission list.

The Remove button on a list only closes cases if the list is a submission list - otherwise candidates are just removed from the list.

Candidate drop down in a list only closes case if submission list.

Update Cases only appears for submission list